Add missing component type
