+++ types/schema/components/index.ts
@@ -4,3 +4,4 @@ export type * from './component-heading.js';
 export type * from './component-media.js';
 export type * from './component-metrics.js';
 export type * from './component-quote.js';
+export type * from './component.js';
